Issue 003: When deleting storage def, also delete backup info.
When user deletes a storage definition via 'atbu creds delete-storage-def',
the storage def's backup information remains present out of an abundance
of caution. This may or may not be convenient.

Provider user options to explicitly request deletion of backup information
while also prompting the user if they are not explicit about it.