Two pipeline stages can communicate via a pipe.
Allow two adjacent pipeline stages to communicate via a pipe, allowing
for a later stage to consume the results of a prior stage immediately,
instead of waiting for the prior stage to fully complete. This reduces
latency and allows for avoiding use of temp files.

Dual stage w/pipe is currently employed for backup compression, where
the compression stage writes compressed file data to the file backup
stage which can immediately begin encryption/upload without having to
wait for the compression stage to fully complete. As well, no temp
files are required for backup compression.