Propagate deduplication_option for unchanged file.
When a file is detected as unchanged based on path/size/date check,
propagate the deduplication_option from the basis file information
object because such maintains the chain to the physical file when
file information is loaded. This is because files with
deduplication_option set are resolved differently at the time backup
information is loaded.

The issue pertaining to this fix can be observed by mixing use of
deduplcation with non-deduplication. For example, backup newly created
duplicates using a dedup option, followed by use of either incremental
or incremental-plus (without any changes is fine). The latest backup
will produce file information that is unchanged without a
deduplication_option, but with a path that cannot be matched to an
existing physical file in the history. This leads to incorrect backup
info load resolution path on load which cannot resolve a backing file
info based on path. By propagating the deduplication_option, the dedup
resolution avenue to be taken, which is digest-driven, which then leads
to appropriately finding the backing file info.