Allow GCS config without a project.
Allow a GCS storage definition configuration to omit a
Google project ID, which means x-goog-project-id will not be set/sent
with respective requests.

Note, if you do not specify a project ID, you may experience a failure
due to the GCS API requiring x-goog-project-id. For example, for a case
where you do not have a default project ID, or the default is not
applicable to the request.