Ignore 'unsuccessful' file info when indexing.
Issue: Backup information saved for each backup currently includes file
information for files which could not be backed up for any reason. If a
file backup fails, the file's information is therefore still saved as
part of the resulting backup information and is marked as unsuccessful.
When loading backup information, build_indexes was not ignoring files
marked as unsuccessful and was therefore attempting to restore files
that had not been backed up.

Fix: During build_indexes, ignore unsuccessful file information.