Ignore drive when finding filesystem storage def
When searching a configuration file for a filesystem storage definition,
ignore the container drive letter as it is the original drive letter and
could change.

As part of this, if the config, found via path to filesystem backup,
only contains a single definition (most always, this will be the case),
simply return that one and only definition.